requirements.txt
Flask
psycopg2-binary
numpy
pandas
matplotlib
scikit-learn